Refactor duplicate code in the inference and the serving code
WIP (#1)

* Moved duplicate pre=processing and post-processing code from the inference and the serving code into `tacotron/input/helpers.py`.
* Refactored the inference and the serving code.
